Enable build flags (#16)
* Added build flags per team process
* Addressed some build fixes found by flags
* Fix build issues with new build flags

+++ cloudwatch_logs_common/include/cloudwatch_logs_common/utils/observable_object.h
@@ -90,7 +90,7 @@ public:
      * Get the current number of listeners
      * @return
      */
-    virtual int getNumberOfListeners() {
+    virtual size_t getNumberOfListeners() {
       return listeners_.size();
     }
